backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
tensorflow==2.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.19
joblib==1.4.2
